break up lib.rs
