+++ crates/btc-notify/src/constants.rs
@@ -1,0 +1,5 @@
+//! This module provides the constant values used throughout the crate.
+
+// TODO(proofofkeags), TODO(Rajil1213): Use different default finality depths depending on the
+// network we are on?
+pub(crate) const DEFAULT_BURY_DEPTH: usize = 6;
